Fixed to make the PersistMetadataXMLTask thread safe by removing use of DocMetaDataXMLParser as a Spring bean and instead using it as a local variable in the DocMetadataServiceImpl.  Also changed type of jobInstanceId from Integer to Long to conform to the Spring Batch API.
git-tfs-id: [https://tfsnpt.int.thomson.com:443/tfs/Cobalt_Collection/]$/Cobalt eBook Builder/Development;C210638

+++ ebookGenerator/src/com/thomsonreuters/uscl/ereader/assemble/service/TitleMetadataService.java
@@ -33,5 +33,5 @@ public interface TitleMetadataService {
 	 * @param titleManifest the title manifest (title.xml) to create.
 	 * @param tocXml the TOC structure from which the &lt;toc&gt; &amp; &lt;docs&gt; portions of the manifest are to be derived.
 	 */
-	public void generateTitleManifest(final OutputStream titleManifest, final InputStream tocXml, final TitleMetadata titleMetadata, final Integer jobInstanceId, final File documentsDirectory);
+	public void generateTitleManifest(final OutputStream titleManifest, final InputStream tocXml, final TitleMetadata titleMetadata, final Long jobInstanceId, final File documentsDirectory);
 }
